Portfolio

+++ app/Http/Controllers/Backend/RoleController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers\Backend;
 
 use App\Http\Controllers\Controller;
 use App\Models\Role;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -15,6 +17,8 @@ class RoleController extends Controller
 
     public function index()
     {
+        abort_if(Gate::denies('role-index'), Response::HTTP_FORBIDDEN, 'Forbidden');
+
         $roles =  Role::orderBy('created_at', 'desc')->get();
         $role = new Role;
 
@@ -43,6 +47,8 @@ class RoleController extends Controller
 
     public function edit(Role $role)
     {
+        abort_if(Gate::denies('role-edit'), Response::HTTP_FORBIDDEN, 'Forbidden');
+
         return view('admin.roles.edit', [
             'role' => $role,
             'submit' => 'Update'
@@ -71,6 +77,8 @@ class RoleController extends Controller
 
     public function destroy($id)
     {
+        abort_if(Gate::denies('role-destroy'), Response::HTTP_FORBIDDEN, 'Forbidden');
+
         $role = Role::find($id);
         $role->delete();
 
@@ -82,6 +90,8 @@ class RoleController extends Controller
 
     public function trash()
     {
+        abort_if(Gate::denies('role-trash'), Response::HTTP_FORBIDDEN, 'Forbidden');
+
         $roles = Role::onlyTrashed()->get();
         return view('admin.roles.trash', compact('roles'));
     }
@@ -89,6 +99,8 @@ class RoleController extends Controller
 
     public function restore($id = null)
     {
+        abort_if(Gate::denies('role-restore'), Response::HTTP_FORBIDDEN, 'Forbidden');
+
         if($id !=null) {
             $id = Role::onlyTrashed()->where('id' , $id)->restore();
         } else {
@@ -102,6 +114,8 @@ class RoleController extends Controller
 
     public function delete($id = null)
     {
+        abort_if(Gate::denies('role-delete'), Response::HTTP_FORBIDDEN, 'Forbidden');
+
         if($id !=null) {
             $id = Role::onlyTrashed()->where('id' , $id)->forceDelete();
         } else {
